requirements.txt
annotated-types==0.7.0
anyio==4.9.0
autogen-agentchat==0.7.1
autogen-core==0.7.1
certifi==2025.7.14
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jsonref==1.1.0
openai==1.98.0
opentelemetry-api==1.36.0
pillow==11.3.0
protobuf==5.29.5
pyautogen==0.10.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
zipp==3.23.0
